fix(text): B1c autofit 多字体 65 档矩阵收敛——CJK/Latin/Myanmar 全绿
B1c 收口：引擎 autohint 对 FreeType 2.11.1 autofit 的逐字形对照，
wqy-nohint/microhei × cjk3000、FreeSans/DejaVuSans × latin_all、
PadaukBook × mymr 全部 65 档（8–72px）四维度（点数/坐标/on-off/分组）
一致；gap 字体（deva/gujr/arab/ethi）剩余坏字待脚本类实现（见下）。

- fix(autohint): ContourPoint X/Y int16→int32 全链路——搬 U+642C
  @70/71/72px 26.6 溢出（fy=15105，33042>32767）；坐标增量保持
  int32(int16(...)) 有符号语义，防花 U+82B1 pt5 符号翻转
- fix(autohint): linkSegmentsCJK serif distThreshold 补 FT_DivFix
  舍入（+scale16/2）——虫@45px 137 vs 136 截断，过滤差异致 122 坏字
- fix(autohint): computeEdges 第一遍按 FT 边表序（插入即排序）扫描，
  非检测序——凰 U+51F0 pos=8 双线时无链接 serif 段归属边不同，
  strict dist<best 取表序靠前边；microhei 65 档 25 坏字清零
- fix(autohint): 蓝区 shoot.fit 离散化 delta 定点化——FT_MulFix
  (ref-shoot)×scale 区高 32 vs float 31，0.5px 档翻转致 dejavu @37px
  顶边吸附 1728 vs 1760（82 坏字）+ freesans @63px 西里尔（24 坏字）
- fix(autohint): 段宽阈值 FT_DivFix(32,scale) 定点——6.91 vs 7，
  $@72px delta=7 段被误滤丢边
- fix(autohint): edt 阈值 Default 分支 FT 两步定点（FT_MulFix→封顶
  0.25px→FT_DivFix），14→15——σ/χ@8px 458 段并入 444 边
- fix(autohint): x-height 校正 scale16dot16 定点重写、锚定 stem 的
  edge2 不标 done（aflatin.c:3228-3236）、interpolateEdge 四舍五入
  （FT_MulDiv 语义）——д@70px stem 漂移等
- fix(autohint): autoHintMetricsKey/glyphScriptKey 加 fontDigest
  （FNV-1a，sync.Map 记忆化）——wqy-microhei 与 nohint 版 FullName
  +UPM 碰撞致缓存串字，microhei 12px 498 坏字清零
- test: zz_b1_ttf_scan_test.go——B1c 65 档矩阵（Probe 12px 探路 +
  CJK 65×2 + Latin 65 + Myanmar 65）；TestB1TTFMyanmar wantCount
  0→b1cSampleCount 修正
- fix(autohint): computeEdgeDistThreshold legacy 路径保留 float 除法
  ——合成轴单元测试（scale=1 无 scale16dot16）不再被 int32 截断破坏

遗留：deva/gujr/ethi/arab 缺脚本类（af_indic 移植）；arab ف 的 IUP
u/v 链边界差异（引擎链自引用 vs FT 跨轮廓回绕）各 1–13 坏字 @12px。

+++ render/text/glyf_parser.go
@@ -27,7 +27,7 @@ import (
 // ContourPoint represents a raw TrueType glyph contour point.
 // Coordinates are in font units (unscaled design space).
 type ContourPoint struct {
-	X, Y    int16 // coordinates in font units (unscaled)
+	X, Y    int32 // coordinates in font units (unscaled)
 	OnCurve bool  // true = on-curve point, false = off-curve control point
 }
 
@@ -221,8 +221,8 @@ func extractGlyfContourOwn(glyfData, locaData []byte, glyphIndex int, isLong boo
 	}
 
 	// Parse X coordinates.
-	xs := make([]int16, numPoints)
-	var prevX int16
+	xs := make([]int32, numPoints)
+	var prevX int32
 	for i := range numPoints {
 		f := flags[i]
 		xShort := f&0x02 != 0
@@ -231,7 +231,7 @@ func extractGlyfContourOwn(glyfData, locaData []byte, glyphIndex int, isLong boo
 			if pos >= len(data) {
 				return nil, fmt.Errorf("text: glyf parser: glyph %d: X coord overflow", glyphIndex)
 			}
-			val := int16(data[pos])
+			val := int32(data[pos])
 			pos++
 			if !xSame {
 				val = -val
@@ -241,7 +241,7 @@ func extractGlyfContourOwn(glyfData, locaData []byte, glyphIndex int, isLong boo
 			if pos+2 > len(data) {
 				return nil, fmt.Errorf("text: glyf parser: glyph %d: X coord overflow", glyphIndex)
 			}
-			prevX += int16(binary.BigEndian.Uint16(data[pos : pos+2]))
+			prevX += int32(int16(binary.BigEndian.Uint16(data[pos : pos+2])))
 			pos += 2
 		}
 		// else: xSame && !xShort → same as previous
@@ -249,8 +249,8 @@ func extractGlyfContourOwn(glyfData, locaData []byte, glyphIndex int, isLong boo
 	}
 
 	// Parse Y coordinates.
-	ys := make([]int16, numPoints)
-	var prevY int16
+	ys := make([]int32, numPoints)
+	var prevY int32
 	for i := range numPoints {
 		f := flags[i]
 		yShort := f&0x04 != 0
@@ -259,7 +259,7 @@ func extractGlyfContourOwn(glyfData, locaData []byte, glyphIndex int, isLong boo
 			if pos >= len(data) {
 				return nil, fmt.Errorf("text: glyf parser: glyph %d: Y coord overflow", glyphIndex)
 			}
-			val := int16(data[pos])
+			val := int32(data[pos])
 			pos++
 			if !ySame {
 				val = -val
@@ -269,7 +269,7 @@ func extractGlyfContourOwn(glyfData, locaData []byte, glyphIndex int, isLong boo
 			if pos+2 > len(data) {
 				return nil, fmt.Errorf("text: glyf parser: glyph %d: Y coord overflow", glyphIndex)
 			}
-			prevY += int16(binary.BigEndian.Uint16(data[pos : pos+2]))
+			prevY += int32(int16(binary.BigEndian.Uint16(data[pos : pos+2])))
 			pos += 2
 		}
 		ys[i] = prevY
@@ -532,8 +532,8 @@ func extractCompositeContoursGuarded(
 			y += comp.dy
 
 			allPoints = append(allPoints, ContourPoint{
-				X:       int16(x),
-				Y:       int16(y),
+				X:       int32(x),
+				Y:       int32(y),
 				OnCurve: pt.OnCurve,
 			})
 		}
